requirements.txt
customtkinter
pywin32
requests
pywifi
comtypes
winreg
pyobjc